Big card refactor
